Modify ProjectController.cs so that the user that belong in the project can modify it or delete it

+++ PDBT/Controllers/ProjectController.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
@@ -62,23 +63,23 @@ namespace PDBT.Controllers
                 return BadRequest();
             }
 
-            var issue = DtoToProject(projectDto);
+            var project = DtoToProject(projectDto);
             
-            _context.Projects.Update(issue);
-            
-            
-            //TODO Implement Project specific labels
-            /*if (projectDto.Labels != null)
-            {
-                // Need to retrive list of current labels to prevent duplicate entries
-                issue = await _context.Issues.GetByIdAsync(id);
+            _context.Projects.Update(project);
+            project = await _context.Projects.GetByIdAsync(project.Id);
 
-                await InsertLabels(issue, issueDto.Labels);
-            }*/
+            //TODO Implement Project specific labels
 
             try
             {
-                await _context.CompleteAsync();
+                if (await verifyUser(project))
+                {
+                    await _context.CompleteAsync();
+                }
+                else
+                {
+                    return Forbid();
+                }
             }
             catch (DbUpdateConcurrencyException)
             {
@@ -106,19 +107,22 @@ namespace PDBT.Controllers
             }
 
             var project = DtoToProject(projectDto);
-          
+            if (User?.Identity?.Name != null)
+            {
+                var authUserId = Int32.Parse(User.Identity.Name);
+                var authUser = await _context.Users.GetByIdAsync(authUserId);
+
+                // A new project will always have the user section as null so we don't need check if it is null
+                project.Users = new List<User>();
+                
+                project.Users.Add(authUser);
+            }
+
+
             _context.Projects.Add(project);
 
             //TODO Implement project specific labels
-            /*if (projectDto.Labels != null)
-            {
-                //Prevents a null reference exception when adding the labels
-                project.Labels = new List<Label>();
-            
-                project = await InsertLabels(project, projectDto.Labels);
-            }*/
 
-            
             await _context.CompleteAsync();
 
             return CreatedAtAction("GetProject", new { id = project.Id }, project);
@@ -139,10 +143,15 @@ namespace PDBT.Controllers
                 return NotFound();
             }
 
-            _context.Projects.Remove(project);
-            await _context.CompleteAsync();
+            if (await verifyUser(project))
+            {
+                _context.Projects.Remove(project);
+                await _context.CompleteAsync();
+                
+                return NoContent();
+            }
 
-            return NoContent();
+            return Forbid();
         }
 
         private Project DtoToProject(ProjectDTO projectDto) =>
@@ -152,6 +161,21 @@ namespace PDBT.Controllers
                 Name = projectDto.Name,
                 Description = projectDto.Description
             };
+
+        private async Task<bool> verifyUser(Project project)
+        {
+            if (User?.Identity?.Name != null)
+            {
+                var authUserId = Int32.Parse(User.Identity.Name);
+                var authUser = await _context.Users.GetByIdAsync(authUserId);
+
+                if (project.Users.Contains(authUser))
+                {
+                    return true;
+                }
+            }
+            return false;
+        }
         
         private bool ProjectExists(int id)
         {
